+++ packages/implementations/react/src/atomic/Tooltip/src/tooltip.stories.tsx
@@ -4,6 +4,7 @@ import { argTypesConvert } from ".storybook/utils";
 import { Box } from "@rarui-react/box";
 import { Text } from "@rarui-react/text";
 import { Button } from "@rarui-react/button";
+import { Title } from "@rarui-react/title";
 
 import { Tooltip } from "./Tooltip";
 import docs from "./tooltip.docs.json";
@@ -12,7 +13,7 @@ const meta = {
   title: "Atomic/Tooltip",
   component: Tooltip,
   render: ({ onVisibility: _onVisibility, ...args }) => (
-    <Box display="flex" justifyContent="center">
+    <Box cursor="default" display="flex" justifyContent="center">
       <Tooltip {...args}>
         <Text color="$primary">Hover</Text>
       </Tooltip>
@@ -31,7 +32,11 @@ type Story = StoryObj<typeof Tooltip>;
 export const basic: Story = {
   args: {
     enabledHover: true,
-    content: <Text color="$primary">Lorem ipsum dolor site amet</Text>,
+    content: (
+      <Text fontSize="$s" color="$primary">
+        Lorem ipsum dolor site amet
+      </Text>
+    ),
   },
 };
 
@@ -48,21 +53,24 @@ export const rich: Story = {
         <Tooltip.Body>
           <Box
             display="flex"
+            flexDirection="column"
             alignItems="center"
             justifyContent="center"
             borderStyle="dashed"
             borderWidth="$1"
             borderColor="$warning"
+            borderRadius="$2xs"
+            gap="$3xs"
             width="100%"
-            height="150px"
+            minHeight="252px"
+            minWidth="260px"
+            backgroundColor="$warning-subdued"
           >
-            <Text
-              textAlign="center"
-              color="$warning"
-              fontSize="$s"
-              lineHeight="$m"
-            >
-              Replace me with your content
+            <Title as="h6" textAlign="center" color="$warning">
+              ❖ Slot
+            </Title>
+            <Text fontSize="$s" textAlign="center" color="$warning">
+              Substitua por Componente
             </Text>
           </Box>
         </Tooltip.Body>