+++ packages/implementations/react/src/composite/Sidebar/src/sidebar.stories.tsx
@@ -4,6 +4,7 @@ import { useArgs } from "@storybook/preview-api";
 import { Box } from "@rarui-react/box";
 import { Button } from "@rarui-react/button";
 import { Text } from "@rarui-react/text";
+import { Title } from "@rarui-react/title";
 import { argTypesConvert } from ".storybook/utils";
 
 import { Sidebar } from "./Sidebar";
@@ -37,15 +38,26 @@ export const basic: Story = {
     children: (
       <Box
         display="flex"
+        flexDirection="column"
         alignItems="center"
         justifyContent="center"
         borderStyle="dashed"
         borderWidth="$1"
         borderColor="$warning"
+        borderRadius="$2xs"
+        gap="$3xs"
         width="100%"
         height="100%"
+        minHeight="252px"
+        minWidth="260px"
+        backgroundColor="$warning-subdued"
       >
-        <Text textAlign="center">Replace me with your content</Text>
+        <Title as="h6" textAlign="center" color="$warning">
+          ❖ Slot
+        </Title>
+        <Text fontSize="$s" textAlign="center" color="$warning">
+          Substitua por Componente
+        </Text>
       </Box>
     ),
   },